Add test for d_logger added (only for unit test)
